refactor: Move `TaskPageModel` into a separate file
